Add ISNMenuValidation to ActionDispatcher
When a user event occurs, it looks at all of the menu items and determines if they have targets set.
Any menu item with a target set checks to see if the target has the action method in question
If the target has the action, it then looks for the validateMenuItem on the target
The way we hadle the event handler version is to create an ActionDispatcher, with the action method set on that object and the dispatcher set as the target.  As a result, any ValidateMenuitem method set in another class while trying to use the ActionDispatcher is never seen.  The method needs to be on the ActionDispatcher itself.
I fixed it by adding a new ValidateMenuItem func to NSMenuItem.  I added the INSMenuValidation interface to the ActionDispatcher with the correct export, then a Func that can be set.  If the function isn't null, it calls it then returns the value.  Otherwise it just returns true.  Then added a ValidateMenuItem Func to NSMenuItem that sets/gets the one from the ActionDispatcher.
Using this build, I was able to set the Activated method and ValidateMenuItem func on a menu item and successfully validate it.