[monotouch-test] Tweak CalendarTest yet again.
Hopefully third time's the charm...

Don't do date math (adding hours) to a local datetime, since DST *will* muddy
the waters and prove that 1=2.

Instead convert the date we want to calculate on to UTC, which should be DST-agnostic.

I've tested this by running the test for every hour during the next 10 years,
so that should cover mostly everything (although I'm still waiting for the
Delorean I ordered to be able to test both in the future and the past).

Previous attempts:

https://github.com/rolfbjarne/xamarin-macios/commit/0442cdf9c0c4cbb3ae2a0f21125dcc98ae85b648
https://github.com/rolfbjarne/xamarin-macios/commit/5caddb35710482574eafd7b9a0cbb8c4aa97d4db