[linker] Refactor and simplify code optimizations.
We had several types of code optimizations, where we'd determine a particular
expression is a constant, then remove the condition and the code for any
dead branches.

Unfortunately each type of expression had its own logic to determine the
sequence of code to remove, each slightly different depending on the actual
conditional expression. This has led to bugs in the past, either when
switching between compilers (mcs/csc), or compilers have changed their emitted
code.

So implement a more generic version that given a branch instruction is able to
determine the exact code sequences that correspond to the true and false
blocks of the condition. This version is also a lot more defensive than the
previous code, it will bail out without doing anything if it finds a code
sequence it doesn't understand.

This also makes it easier to implement other similar code optimizations.