[SceneKit] Add `AddAnimation` ovearload that takes a `ISCNAnimationProtocol`
Fixes xamarin/xamarin-macios#3166

Apple broke `addAnimation:forKey:` API by changing `CAAnimation` parameter into a
`ISCNAnimationProtocol` that was introduced in Xcode 9 (iOS 11 and company).

We can't break the existing API but we can add an overload that
reuses the same selector with an sligthly twist in order to avoid
API ambiguity (because `CAAnimation` conforms to `SCNAnimation`) the
`key` parameter is being exposed as `string` and also being fixed
in XAMCORE_4_0 this way. `key` being a string instead of a `NSSrring`
makes sense because the `key` is just an identifier for the animation
and can be set to null so having a NET string vs a `NSString` should
make using this API a lot easier from a NET context.

The `new void AddAnimation` is used to force the generator
to generate (^.^) the `CAAnimation, NSString` overload in the
objects that implement `ISCNAnimationProtocol` else you would get
a CS0535.