[runtime] Add a few asserts to avoid reading memory.
We're looping over a char* array in multiple places, not checking whether we'd
reach the end of the string.

These asserts make sure that we don't looping over invalid memory when we find
something we don't understand.

If given random invalid memory, we'd eventually either crash or hit an assert
further below, so these new asserts won't cause any change in behavior, but it
improves the diagnostic message.

I ran into this with the test case from bug #[61086][1].

[1]: https://bugzilla.xamarin.com/show_bug.cgi?id=61086