[generator] Ensure we don't generate code that creates pointers to local variables that then go out of scope.
Instead of generating the following:

    var converted = IntPtr.Zero;
    if (value.HasValue) {
    	var v = value.Value;
    	converted = new IntPtr (&v);
    }
    // now 'converted' points to a local variable that is out of scope

do this:

    var converted = IntPtr.Zero;
    var v = default (float);
    if (value.HasValue) {
    	v = value.Value;
    	converted = new IntPtr (&v);
    }
    // now 'converted' points to a local variable that's still in scope

so that we don't store a pointer to a local variable and use that pointer
after the local variable has gone out of scope.

I'm not entirely sure this really is a problem in C#, but it doesn't hurt to
be on the safe side.

Also add newlines to the generated code to make it look nicer.

Generator diff: https://gist.github.com/rolfbjarne/225d920b7fa28d07fbeab13754cebd2a