[src] Fix installing pdbs for MonoTouch.Dialog and MonoTouch.NUnitLite.
The pdb rules I've removed here didn't do anything (no recipe, only a
dependency, which meant that the rule to install them wouldn't execute because
of some reason only make understands).

Instead implement this correctly: one command (running the C# compiler)
produces both the dll and the pdb, and the way to express this in a makefile
is to use a pattern rule [1] with two targets (the dll and the pdb), and both
targets having with the same stem (it can be anything, I chose the path
separator because that was simplest, even though it looks somewhat strange).

[1]: https://www.gnu.org/software/make/manual/html_node/Pattern-Examples.html#Pattern-Examples