[src] Make sure that the output is always a dll.
In the previous commit I changed these rules to state that the same recipe
would build both the dll and the pdb.

A problem arises when the file that ends up executing the recipe is the pdb,
in which case we can't say that the output from the compilation is the target
(`-out:$@`), because that would write the assembly with the pdb's filename.

So change the output to be a dll instead.