[msbuild] Use a response file for references in MTouchTaskBase
The cause of the warning: "warning MSB6002: The command-line for the 'MTouch' task is too long"
was the number of references that can be passed to `mtouch`. Now all those references (-r) are written in
a response file which in turn is passed to mtouch.

- Fixes bug #56501: MSB6002 command-line for MTouch task is too long, > 32000 characters
(https://bugzilla.xamarin.com/show_bug.cgi?id=56501)
- The response file is created in the device sepcific folder, as `response-file.rsp`. It is re-created every time.
- Added msbuild test to ensure the response file is created by `GenerateCommandLineCommands` and that
  it includes all the references.